test: LinkMessage
