+++ liquid/src/main/java/foundation/algorand/auth/connect/SignalInterface.kt
@@ -4,31 +4,10 @@ import android.content.Context
 import android.graphics.Bitmap
 import okhttp3.OkHttpClient
 import io.socket.client.Socket
-import org.json.JSONObject
 import org.webrtc.PeerConnection
 import org.webrtc.DataChannel
 import org.webrtc.SessionDescription
 import kotlin.math.floor
-
-class LinkMessage(val requestId: Double, val wallet: String, val credId: String? = null) {
-    companion object {
-        const val TAG = "connect.LinkMessage"
-        fun fromJson(json: String): LinkMessage {
-            val data = JSONObject(json).get("data") as JSONObject
-            val requestId = data.get("requestId").toString().toDouble()
-            val wallet = data.get("wallet").toString()
-            val credId = data.get("credId").toString()
-            return LinkMessage(requestId, wallet, credId)
-        }
-    }
-    fun toJson(): JSONObject {
-        val result = JSONObject()
-        result.put("requestId", requestId)
-        result.put("wallet", wallet)
-        result.put("credId", credId)
-        return result
-    }
-}
 
 interface SignalInterface {
     val url: String // URL of Signal Server
